introduce abstract class and refactor
