+++ src/view/films-container-view.js
@@ -1,23 +1,9 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createFilmsContainerTemplate = () => '<div class="films-list__container"></div>';
 
-export default class FilmsContainerView {
-  #element = null;
-
+export default class FilmsContainerView extends AbstractView {
   get template() {
     return createFilmsContainerTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
