+++ src/view/films-list-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createFilmsListTemplate = () => (
   `<section class="films">
@@ -8,22 +8,8 @@ const createFilmsListTemplate = () => (
   </section>`
 );
 
-export default class FilmsListView {
-  #element = null;
-
+export default class FilmsListView extends AbstractView {
   get template() {
     return createFilmsListTemplate();
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
