+++ src/view/popup-comment-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 import { humanizeCommentDate } from '../util.js';
 
 const createPopupCommentTemplate = (commentary) => {
@@ -22,27 +22,15 @@ const createPopupCommentTemplate = (commentary) => {
   );
 };
 
-export default class PopupCommentView {
-  #element = null;
+export default class PopupCommentView extends AbstractView {
   #comment = null;
 
   constructor(comment) {
+    super();
     this.#comment = comment;
   }
 
   get template() {
     return createPopupCommentTemplate(this.#comment);
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
