+++ src/view/popup-genre-view.js
@@ -1,28 +1,16 @@
-import { createElement } from '../render.js';
+import AbstractView from '../framework/view/abstract-view.js';
 
 const createPopupGenreTemplate = (genre) => `<span class="film-details__genre">${genre}</span>`;
 
-export default class PopupGenreView {
-  #element = null;
+export default class PopupGenreView extends AbstractView {
   #genre = null;
 
   constructor(genre) {
+    super();
     this.#genre = genre;
   }
 
   get template() {
     return createPopupGenreTemplate(this.#genre);
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
